refactor/restructure-components: create components folder structure
