+++ src/Inspire/Inspire.tsx
@@ -1,20 +1,20 @@
 import React, { useMemo } from "react";
 import { animated, useSpring } from "react-spring";
 import styled from "styled-components/macro";
-import Picture from "../Style/Picture";
+import Picture from "../Components/Picture/Picture";
 import {
   Body,
   Header,
   headerSpringProps,
   StyledContainer,
-} from "../Style/Stuff";
-import Back from "../Style/StyledBack";
+} from "../Components/Stuff/Stuff";
+import Back from "../Components/Button/StyledBack";
 import { Button } from "@material-ui/core";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../store/store";
 import { fetchQuote } from "../store/spring";
 import { useWindowSize } from "@react-hook/window-size";
-import breakpoint from "../Style/Common/breakpoints";
+import breakpoint from "../Components/Common/breakpoints";
 
 export function Inspire(): JSX.Element {
   const headerProps = useSpring(headerSpringProps);
